FIN: EasyTransfer for finger

+++ Gripper_Arduino/main_board/lights.cpp
@@ -4,8 +4,8 @@
 void Lights::tick(){
   switch(_state){
     case CALIBRATING:{
-      for (int j=0;j<4;j++){
-        for (int i = 0; i < 8; i++){
+      for (byte j=0;j<4;j++){
+        for (byte i = 0; i < 8; i++){
           digitalWrite(_pins[i], HIGH);
           delay(25);
           digitalWrite(_pins[i], LOW);
@@ -14,23 +14,23 @@ void Lights::tick(){
       _state = READY;
     }break;
     case READY:{
-      for (int i = 0; i < 8; i++){
+      for (byte i = 0; i < 8; i++){
         digitalWrite(_pins[i], LOW);
       }
     }break;
   }
 }
 
-void Lights::setLight(int pin){
-  for (int i = 0; i < 8; i++){
+void Lights::setLight(byte pin){
+  for (byte i = 0; i < 8; i++){
     digitalWrite(_pins[i], LOW);
   }
   digitalWrite(_pins[pin],HIGH);
   _state = SET;
 }
 
-void Lights::setLight(int pina, int pinb){
-  for (int i = 0; i < 8; i++){
+void Lights::setLight(byte pina, byte pinb){
+  for (byte i = 0; i < 8; i++){
     digitalWrite(_pins[i], LOW);
   }
   digitalWrite(_pins[pina],HIGH);
